// src/interfaces/bookingInterface.ts
export interface BookingInterface {
    _id?: string;
    guest: string;
    phoneNumber: string;
    orderDate: string;
    checkIn: string;
    checkOut: string;
    specialRequest: string;
    roomId: String;
    roomType: string;
    roomNumber: string;
    status: string;
    photos: string[];
}